POC_mcp_testscript/test_mcp_orchestrator/pyproject.toml
[project]
name = "mcp-web-auto"
version = "0.1.0"
description = ""
authors = ["kimtth <kimtth@email.com>"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.8.4",
    "fastmcp>=0.4.1",
    "loguru>=0.6.0",
    "mcp>=1.1.1",
    "openai>=1.57.1",
    "playwright>=1.55.0",
    "pydantic>=2.10.3",
    "pytest-playwright>=0.6.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.24.0",
    "typing-extensions>=4.12.2",
    "watchdog>=3.0.0",
]

